+++ leda/components/DropZone/handlers.ts
@@ -1,14 +1,21 @@
 import * as React from 'react';
 import { DropzoneRef } from 'react-dropzone';
 import { isFunction } from 'lodash';
-import { checkFiles } from './helpers';
+
+import { checkFiles, removeFromFileList, addToFileList } from './helpers';
 import {
-  ChangeEventHandler, DropZoneProps, DropZoneState, FileType,
+  ChangeEventHandler, DropZoneProps, DropZoneState,
 } from './types';
 
+/**
+ * Function creates click event handler
+ * @param {DropZoneProps} props - properties of DropZone component
+ * @param {React.MutableRefObject<DropzoneRef | null>} ref - ref of DropZone component
+ *
+ * @returns {React.MouseEventHandler => (ev: React.MouseEvent<HTMLDivElement>): void} - click event handler
+ */
 export const createClickHandler = (
   { onClick, isDisabled }: DropZoneProps,
-  state: DropZoneState,
   ref: React.MutableRefObject<DropzoneRef | null>,
 ): React.MouseEventHandler => (ev: React.MouseEvent<HTMLDivElement>): void => {
   if (!isDisabled && ref.current && (ev.target as HTMLDivElement).tagName === 'BUTTON') ref.current.open();
@@ -16,7 +23,14 @@ export const createClickHandler = (
   if (isFunction(onClick)) onClick(ev);
 };
 
-
+/**
+ * Function creates change event handler
+ * @param {DropZoneProps} props - properties of DropZone component
+ * @param {DropZoneState} state - state of DropZone component
+ * @param {React.Dispatch<React.SetStateAction<DropZoneState>>} setState - setState function of DropZone component
+ *
+ * @returns {ChangeEventHandler} - change event handler
+ */
 export const createChangeHandler = (
   props: DropZoneProps,
   state: DropZoneState,
@@ -31,36 +45,29 @@ export const createChangeHandler = (
 
   const value = valueProp || state;
 
-  // В обработчик передаем только те принятые файлы, которые еще не были добавлены
-  const [acceptedFiles, rejectedFiles] = !removedFile ? checkFiles(props, state, accepted, rejected) : [accepted, rejected];
+  const [acceptedFiles, rejectedFiles] = checkFiles(props, state, accepted, rejected);
 
-  const acceptedWithRemoved = removedFile ? (value.acceptedFiles as FileType[]).filter((file: FileType): boolean => file !== removedFile) : [...acceptedFiles, ...value.acceptedFiles];
-
-  const newValue = {
-    acceptedFiles: acceptedWithRemoved,
-    rejectedFiles,
-  };
+  const { newDropped, newValue } = removedFile
+    ? removeFromFileList(removedFile, rejected, value.acceptedFiles)
+    : addToFileList(acceptedFiles, rejectedFiles, value.acceptedFiles);
 
   const customEvent = {
     ...ev,
     component: {
       ...ev?.target,
-      dropped: {
-        acceptedFiles: accepted,
-        rejectedFiles: rejected,
-      },
+      dropped: newDropped,
       value: newValue,
       removedFile,
     },
   };
-  // контролируемый режим
+  // controlled mode
   if (isFunction(onChange)) {
     onChange(customEvent);
   }
 
   if (valueProp) return newValue;
 
-  // неконтролируемый режим
+  // uncontrolled mode
   setState(newValue);
 
   return newValue;